Cleanup AbstractModel test files
- https://github.com/QutEcoacoustics/workbench-client/pull/223#discussion_r415464693
- https://github.com/QutEcoacoustics/workbench-client/pull/223#discussion_r415464945
